4.4 reflections

+++ src/rmath.h
@@ -3,9 +3,9 @@
 #include <math.h>
 #include "rmath_types.h"
 
-RINLINE f32 vec3_dot(vec3 vector_0, vec3 vector_1)
+RINLINE f64 vec3_dot(vec3 vector_0, vec3 vector_1)
 {
-    f32 p = 0;
+    f64 p = 0;
     p += vector_0.x * vector_1.x;
     p += vector_0.y * vector_1.y;
     p += vector_0.z * vector_1.z;
@@ -20,12 +20,12 @@ RINLINE vec3 vec3_sub(vec3 vector_0, vec3 vector_1)
         vector_0.z - vector_1.z};
 }
 
-RINLINE f32 vec3_length_squared(vec3 vector)
+RINLINE f64 vec3_length_squared(vec3 vector)
 {
     return vector.x * vector.x + vector.y * vector.y + vector.z * vector.z;
 }
 
-RINLINE f32 vec3_length(vec3 vector)
+RINLINE f64 vec3_length(vec3 vector)
 {
     return sqrt(vec3_length_squared(vector));
 }
@@ -38,7 +38,7 @@ RINLINE vec3 vec3_add(vec3 vector_0, vec3 vector_1)
         vector_0.z + vector_1.z};
 }
 
-RINLINE vec3 vec3_mul_scalar(vec3 vector_0, f32 scalar)
+RINLINE vec3 vec3_mul_scalar(vec3 vector_0, f64 scalar)
 {
     return (vec3){
         vector_0.x * scalar,
@@ -48,7 +48,7 @@ RINLINE vec3 vec3_mul_scalar(vec3 vector_0, f32 scalar)
 
 RINLINE void vec3_normalize(vec3 *vector)
 {
-    const f32 length = vec3_length(*vector);
+    const f64 length = vec3_length(*vector);
     vector->x /= length;
     vector->y /= length;
     vector->z /= length;
