+++ src/rmath_types.h
@@ -3,36 +3,36 @@
 
 typedef union vec2_u
 {
-    f32 elements[2];
+    f64 elements[2];
     struct
     {
         union
         {
-            f32 x, w;
+            f64 x, w;
         };
         union
         {
-            f32 y, h;
+            f64 y, h;
         };
     };
 } vec2;
 
 typedef union vec3_u
 {
-    f32 elements[3];
+    f64 elements[3];
     struct
     {
         union
         {
-            f32 x, w, r;
+            f64 x, w, r;
         };
         union
         {
-            f32 y, h, g;
+            f64 y, h, g;
         };
         union
         {
-            f32 z, b;
+            f64 z, b;
         };
     };
 } vec3;
